refactor: change component directory structure (#6)
