+++ apps/meetie-service/src/components/MyPage/InformationCard/InformationCard.tsx
