+++ apps/meetie-service/src/components/MyPage/Profile/StudyExperienceCard/StudyExperienceCard.tsx
