+++ apps/meetie-service/src/components/MyPage/dividers/ThickDivider/ThickDivider.tsx
